move address screens to address folder
