+++ lib/main.dart
@@ -9,12 +9,12 @@ import 'package:disappear/screens/challenge/join_challenge_screen.dart';
 import 'package:disappear/screens/chatbot/chatbot_empty_screen.dart';
 import 'package:disappear/screens/chatbot/chatbot_screen.dart';
 import 'package:disappear/screens/chatbot/new_chatbot_screen.dart';
-import 'package:disappear/screens/checkout/add_new_address_screen.dart';
+import 'package:disappear/screens/address/add_new_address_screen.dart';
 import 'package:disappear/screens/checkout/address/checkout_address_screen.dart';
-import 'package:disappear/screens/checkout/address_list_screen.dart';
+import 'package:disappear/screens/address/address_list_screen.dart';
 import 'package:disappear/screens/checkout/checkout_screen.dart';
 import 'package:disappear/screens/checkout/choose_transport_screen.dart';
-import 'package:disappear/screens/checkout/edit_old_address_screen.dart';
+import 'package:disappear/screens/address/edit_old_address_screen.dart';
 import 'package:disappear/screens/checkout/payment_method/ewallet_transfer_screen.dart';
 import 'package:disappear/screens/checkout/payment_method/manual_transfer_screen.dart';
 import 'package:disappear/screens/checkout/use_coupon_screen.dart';
